+++ backend/src/test/java/org/donggle/backend/application/service/MemberServiceTest.java
@@ -2,7 +2,6 @@ package org.donggle.backend.application.service;
 
 import org.donggle.backend.application.repository.MemberCredentialsRepository;
 import org.donggle.backend.application.repository.MemberRepository;
-import org.donggle.backend.application.service.auth.MemberService;
 import org.donggle.backend.domain.member.Member;
 import org.donggle.backend.domain.member.MemberCredentials;
 import org.donggle.backend.domain.member.MemberName;
