+++ backend/src/test/java/org/donggle/backend/application/service/WritingServiceTest.java
@@ -2,7 +2,7 @@ package org.donggle.backend.application.service;
 
 import org.donggle.backend.application.repository.WritingRepository;
 import org.donggle.backend.application.service.request.WritingModifyRequest;
-import org.donggle.backend.application.service.writing.WritingContentService;
+import org.donggle.backend.application.service.writing.WritingFacadeService;
 import org.donggle.backend.domain.writing.Title;
 import org.donggle.backend.domain.writing.Writing;
 import org.donggle.backend.exception.notfound.WritingNotFoundException;
@@ -21,7 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 @SpringBootTest
 class WritingServiceTest {
     @Autowired
-    private WritingContentService writingService;
+    private WritingFacadeService writingService;
     @Autowired
     private WritingRepository writingRepository;
